pyproject.toml
[project]
name = "ecommerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standart]>=0.129.0",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.40.0",
]
